+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove unused `Rpc` trait.
 - Remove `cbor-smol` re-export as `serde`.
 - Make `ctap2::CtapMappingError` private.
+- Set `#[repr(u8)]` for `ctap2::Error` and implement `From<ctap2::Error>` for `u8`.
 
 ## [0.6.0-rc.4] 2026-06-01